refactor(tests): move token view tests into enrollment_littlepay
update module name in dotted paths and imports so that tests are correct
and pass.
